feat: Add dormant YAML report formatter
Introduces a new  function and wires it into the main application entrypoint.

This change replaces the previous Markdown-based report format. The feature is considered 'dormant' because the acceptance tests have not yet been updated to parse the new YAML format, meaning the test suite will fail until the tests are refactored.

This commit also includes:
- A new unit test for the YAML formatter.
- Refactoring of integration tests that were dependent on the old formatter to align with the new YAML output.